requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
html5lib==1.0.1
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
PyJWT==1.6.4
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2018.5
reportlab==3.5.9
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
webencodings==0.5.1
xhtml2pdf==0.2.3
